+++ package.json
@@ -31,7 +31,7 @@
     "lint": "eslint .",
     "lint:fix": "npm run lint -- --fix",
     "test:node": "node --test test/index.cjs",
-    "test:types": "tsd --files 'types/*'",
+    "test:types": "tstyche --tsconfig types/tsconfig.test.json",
     "test": "npm-run-all build test:node test:types",
     "prepublishOnly": "npm run build"
   },
@@ -93,7 +93,7 @@
     "neostandard": "^0.13.0",
     "npm-run-all": "^4.1.5",
     "rimraf": "^6.0.1",
-    "tsd": "^0.33.0",
+    "tstyche": "^7.0.0",
     "typescript": "~6.0.2"
   }
 }
