+++ tstyche.json
@@ -1,0 +1,3 @@
+{
+  "testFileMatch": ["types/**/*.tst.mts"]
+}
